fix: 공통 타입 shared로 분리 (#42)
