+++ app/(dashboard)/my/_api/get-my-account-iamges.ts
@@ -1,20 +1,9 @@
-import { ImageDataModel } from '@/app/(dashboard)/_ui/analysis-container/account-content'
-
 import axiosInstance from '@/shared/api/axios'
-
-interface ResponseModel {
-  content: ImageDataModel
-  first: boolean
-  last: boolean
-  page: number
-  size: number
-  totalElements: number
-  totalPages: number
-}
+import { AccountImageDataModel } from '@/shared/types/strategy-data'
 
 const getMyAccountImages = async (
   strategyId: number
-): Promise<ResponseModel | null | undefined> => {
+): Promise<AccountImageDataModel | null | undefined> => {
   try {
     const response = await axiosInstance.get(`/api/my-strategies/${strategyId}/account-images`)
     return response.data.result
